% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint_tc}
\alias{confint_tc}
\alias{confint.henderson}
\alias{confint.clf}
\alias{confint.robust_henderson}
\title{Confidence Intervals for Trend-Cycle Estimates}
\usage{
\method{confint}{henderson}(object, parm, level = 0.95, asymmetric_var = TRUE, ...)

\method{confint}{clf}(object, parm, level = 0.95, asymmetric_var = TRUE, ...)

\method{confint}{robust_henderson}(object, parm, level = 0.95, asymmetric_var = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{"tc_estimates"}.}

\item{parm}{Unused parameter, kept for compatibility with the generic function.}

\item{level}{The confidence level required (defaults to \code{0.95}).}

\item{asymmetric_var}{Logical. If \code{TRUE} (the default), the variance is estimated
for each asymmetric filter. If \code{FALSE}, the variance associated
with the symmetric filter is used throughout.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{matrix} or \code{mts} object with the filtered series and the lower and upper bounds of the confidence interval.
}
\description{
Computes confidence intervals for trend-cycle estimates of class \code{"tc_estimates"}.
}
\details{
See \code{\link[rjd3filters:confint_filter]{rjd3filters::confint_filter()}} for details on the computation of confidence intervals.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
x <- cars_registrations
tc <- henderson_smoothing(x)
confint <- confint(tc)

plot(confint, plot.type = "single",
	 col = c("red", "black", "black"),
	 lty = c(1, 2, 2), xlab = NULL, ylab = NULL)
lines(x, col = "grey")
legend("topleft", legend = c("x", "Smoothed", "CI (95\%)"),
	   col= c("grey", "red", "black"), lty = c(1, 1, 2))
\dontshow{\}) # examplesIf}
}
