% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_filter.R
\name{henderson_robust_smoothing}
\alias{henderson_robust_smoothing}
\title{Robust Trend-Cycle Estimation using the Henderson Filter}
\usage{
henderson_robust_smoothing(
  x,
  endpoints = c("Musgrave", "QL", "CQ", "DAF"),
  length = NULL,
  ao = NULL,
  ao_tc = NULL,
  ls = NULL,
  icr = NULL,
  local_icr = FALSE,
  asymmetric_var = FALSE,
  degree = 3,
  ...
)
}
\arguments{
\item{x}{An input time series (object of class \code{"ts"} or \code{"mts"}).}

\item{endpoints}{Character string specifying the method used for asymmetric filters.
By default, \code{"musgrave"} is used.}

\item{length}{An integer specifying the length of the symmetric filter.
If \code{NULL}, the length is selected automatically using \code{\link[=x11_trend_selection]{x11_trend_selection()}}.}

\item{ao}{Vector of dates for Additive Outliers (AO) whose effects are allocated
to the irregular component.}

\item{ao_tc}{Vector of dates for Additive Outliers (AO) whose effects are allocated
to the trend-cycle component.}

\item{ls}{Vector of dates for Level Shifts (LS) whose effects are allocated
to the trend-cycle component.}

\item{icr}{Numeric value specifying the Irregular-to-Trend (I/C) ratio used for
asymmetric filters.}

\item{local_icr}{Logical. If \code{TRUE}, the I/C ratio is estimated locally
(as described in Quartier-la-Tente, 2024) instead of globally.}

\item{asymmetric_var}{Logical. When \code{local_icr = TRUE}, if \code{TRUE}, the variance
is estimated for each asymmetric filter. If \code{FALSE} (the default), the variance
associated with symmetric estimates is used throughout.}

\item{degree}{Integer. When \code{local_icr = TRUE}, degree of the polynomial used
to estimate the local bias parameter.}

\item{...}{Additional arguments passed to \code{\link[rjd3filters:lp_filter]{rjd3filters::lp_filter()}}.}
}
\value{
An object of class \code{c("tc_estimates", "robust_henderson")}.
See \code{\link[=tc_estimates]{tc_estimates()}} for a full description of the returned object.
}
\description{
Estimates the trend-cycle component using Henderson moving averages robust to
outliers (additive outliers and level shifts).
}
\details{
When outliers are present in a time series, standard linear filters like the Henderson
moving average can spread their impact across adjacent trend estimates. This function
explicitly accounts for specified Additive Outliers (AO) and Level Shifts (LS) during
the filtering process to prevent distortion of the trend-cycle estimates as described in Quartier-la-Tente (2025).

In a nutshell, Henderson smoothing is equivalent to a local polynomial regression of degree 3
(or equivalently 2) estimated with weighted least squares, using specific weights to obtain
the Henderson coefficients. Adding regressors to this local regression to account for outliers
allows removing their influence on the estimation of polynomial coefficients and thus on the
trend-cycle estimates.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
x <- cars_registrations
outliers <- x13_regarima_outliers(x)
tc <- henderson_robust_smoothing(x, ao = outliers$ao, ls = outliers$ls, local_icr = TRUE)

tc_henderson <- henderson_smoothing(x)
plot(tc, xlim = c(2019, 2022))
lines(tc_henderson$tc, col = "blue")
\dontshow{\}) # examplesIf}
}
\references{
Quartier-la-Tente, A. (2025). Estimation de la tendance-cycle avec des méthodes robustes aux points atypiques. \url{https://github.com/AQLT/robustMA}.
}
