% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/henderson.R
\name{henderson_smoothing}
\alias{henderson_smoothing}
\title{Trend-Cycle Estimation using the Henderson Filter}
\usage{
henderson_smoothing(
  x,
  endpoints = c("Musgrave", "QL", "CQ", "CC", "DAF", "CN"),
  length = NULL,
  icr = NULL,
  local_icr = FALSE,
  asymmetric_var = FALSE,
  degree = 3,
  ...
)
}
\arguments{
\item{x}{An input time series (object of class \code{"ts"} or \code{"mts"}).}

\item{endpoints}{Character string specifying the method used for asymmetric filters.
By default, \code{"musgrave"} is used.}

\item{length}{An integer specifying the length of the symmetric filter.
If \code{NULL}, the length is selected automatically using \code{\link[=x11_trend_selection]{x11_trend_selection()}}.}

\item{icr}{Numeric value specifying the Irregular-to-Trend (I/C) ratio used for
asymmetric filters.}

\item{local_icr}{Logical. If \code{TRUE}, the I/C ratio is estimated locally
(as described in Quartier-la-Tente, 2024) instead of globally.}

\item{asymmetric_var}{Logical. When \code{local_icr = TRUE}, if \code{TRUE}, the variance
is estimated for each asymmetric filter. If \code{FALSE} (the default), the variance
associated with symmetric estimates is used throughout.}

\item{degree}{Integer. When \code{local_icr = TRUE}, degree of the polynomial used
to estimate the local bias parameter.}

\item{...}{Additional arguments passed to \code{\link[rjd3filters:lp_filter]{rjd3filters::lp_filter()}}.}
}
\value{
An object of class \code{c("tc_estimates", "henderson")}.
See \code{\link[=tc_estimates]{tc_estimates()}} for a full description of the returned object.
}
\description{
Estimates the trend-cycle component using Henderson moving averages.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
x <- cars_registrations
tc <- henderson_smoothing(x)
plot(tc, xlim = c(2022, 2025))
\dontshow{\}) # examplesIf}
}
\references{
Henderson, R. (1916).
Note on Graduation by Adjusted Average.
\emph{Transactions of the Actuarial Society of America} 17: 43-48.

Musgrave, J. (1964).
A Set of End Weights to End All End Weights.
\emph{US Census Bureau [Custodian]}.
\url{https://www.census.gov/library/working-papers/1964/adrm/musgrave-01.html}.

Quartier-la-Tente, A. (2024). Improving Real-Time Trend Estimates Using Local Parametrization of Polynomial Regression Filters.
\emph{Journal of Official Statistics, 40}(4), 685-715.
\doi{10.1177/0282423X241283207}.
}
