% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_implicit_forecast.R
\name{implicit_forecasts}
\alias{implicit_forecasts}
\title{Compute Implicit Forecasts}
\usage{
implicit_forecasts(x, ...)
}
\arguments{
\item{x}{a \code{"tc_estimates"} object otherwise uses the \code{\link[rjd3filters:implicit_forecast]{rjd3filters::implicit_forecast()}} function.}

\item{...}{other unused parameters.}
}
\description{
Compute Implicit Forecasts
}
