% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_implicit_forecast.R
\name{implicit_forecasts.henderson}
\alias{implicit_forecasts.henderson}
\alias{implicit_forecasts.clf}
\alias{implicit_forecasts.robust_henderson}
\title{Implicit Forecasts Methodes}
\usage{
\method{implicit_forecasts}{henderson}(x, ...)

\method{implicit_forecasts}{clf}(x, ...)

\method{implicit_forecasts}{robust_henderson}(x, ...)
}
\arguments{
\item{x}{a \code{"tc_estimates"} object otherwise uses the \code{\link[rjd3filters:implicit_forecasts]{rjd3filters::implicit_forecasts()}} function.}

\item{...}{other unused parameters.}
}
\description{
\code{\link[rjd3filters:implicit_forecasts]{rjd3filters::implicit_forecasts()}} methods for \code{"tc_estimates"} objects.
}
\details{
See \code{\link[rjd3filters:implicit_forecasts]{rjd3filters::implicit_forecasts()}} for details on the computation of implicit forecasts.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
tc_mod <- henderson_smoothing(french_ipi[, "manufacturing"])
implicit_forecasts(tc_mod)
\dontshow{\}) # examplesIf}
}
