% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lollypop.R
\name{lollypop}
\alias{lollypop}
\alias{gglollypop}
\title{Lollypop plot}
\usage{
lollypop(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  color_points = col_sa,
  cex_points = 1,
  pch_points = 16,
  xlab = "",
  ylab = "",
  lty_last_tc = 2,
  n_last_tc = 4,
  ...
)

gglollypop(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  color_points = col_sa,
  cex_points = 1,
  pch_points = 16,
  legend_tc = "Trend-cycle",
  legend_sa = "Seasonally adjusted",
  lty_last_tc = 2,
  n_last_tc = 4,
  ...
)
}
\arguments{
\item{object}{\code{"tc_estimates"} object.}

\item{xlim, ylim}{x and y limits of the plot.
If \code{xlim} is defined and not \code{ylim}, then \code{ylim} is determined automatically.}

\item{col_sa, col_tc}{color of the seasonally adjusted and trend-cycle components.}

\item{color_points, cex_points}{color and size of the points associated to the seasonnaly adjusted component.}

\item{pch_points}{point type of the seasonally adjusted component.}

\item{xlab, ylab}{x and y axis labels.}

\item{lty_last_tc}{line type of the last values of the trend-cycle component.}

\item{n_last_tc}{number of last values of the trend-cycle component to be plotted with a different line type
(to emphasize that there is higher variability for the last estimates).
If \code{NULL}, then \code{n_last_tc} is equal to the MCD statistic.}

\item{...}{other parameters.}

\item{legend_tc, legend_sa}{legend of the trend-cycle and seasonally adjusted components.}
}
\value{
\code{lollypop()} returns \code{NULL}, called for side effect.
\code{gglollypop()} returns a \code{ggplot} object.
}
\description{
Lollypop plot
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
tc_mod <- henderson_smoothing(french_ipi[, "manufacturing"])
lollypop(tc_mod, xlim = c(2022, 2024.5))
\dontshow{\}) # examplesIf}
}
