% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tc_estimates}
\alias{plot.tc_estimates}
\alias{lines.tc_estimates}
\alias{autoplot.tc_estimates}
\alias{autoplot.ts}
\title{Plotting Methods for Trend-Cycle Estimates}
\usage{
\method{plot}{tc_estimates}(
  x,
  y = NULL,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  xlab = "",
  ylab = "",
  lty_last_tc = 2,
  n_last_tc = 4,
  ...
)

\method{lines}{tc_estimates}(x, col_tc = "#E69F00", lty_last_tc = 2, n_last_tc = 4, ...)

\method{autoplot}{tc_estimates}(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  legend_tc = "Trend-cycle",
  legend_sa = "Seasonally adjusted",
  lty_last_tc = 2,
  n_last_tc = 4,
  ...
)

\method{autoplot}{ts}(object, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"tc_estimates"}.}

\item{y}{Unused parameter, kept for compatibility with the generic function.}

\item{xlim, ylim}{Limits for the x- and y-axes.
If \code{xlim} is specified and \code{ylim} is \code{NULL}, \code{ylim} is determined automatically
based on the truncated series.}

\item{col_sa, col_tc}{Colors used for the seasonally adjusted and trend-cycle components,
respectively.}

\item{xlab, ylab}{Character strings for the x- and y-axis labels.}

\item{lty_last_tc}{Line type for the last values of the trend-cycle component.}

\item{n_last_tc}{Number of final values of the trend-cycle component to plot
with a distinct line type (\code{lty_last_tc}), emphasizing higher uncertainty in
recent estimates. If \code{NULL} (the default), \code{n_last_tc} is set to the Month of
Cyclical Dominance (MCD) statistic.}

\item{...}{Additional graphical parameters passed to internal plotting functions.}

\item{object}{An object of class \code{"tc_estimates"}.}

\item{legend_tc, legend_sa}{Character strings specifying the legend labels for
the trend-cycle and seasonally adjusted components, respectively.}
}
\value{
\itemize{
\item \code{plot()} and \code{lines()} return \code{NULL} invisibly and are called for their side effect (drawing a plot).
\item \code{autoplot()} returns a \code{\link[ggplot2]{ggplot}} object.
}
}
\description{
Plotting methods for objects of class \code{"tc_estimates"}. \code{plot()} and \code{lines()}
produce base R graphics, whereas \code{autoplot()} creates a \pkg{ggplot2} object.
}
\details{
\code{plot.tc_estimates()} and \code{autoplot.tc_estimates()} produce a plot showing both
the seasonally adjusted series and the estimated trend-cycle component.
\code{lines.tc_estimates()} adds only the trend-cycle component to an existing base R plot.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
tc_mod <- henderson_smoothing(french_ipi[, "manufacturing"])
plot(tc_mod, xlim = c(2022, 2024.5))
\dontshow{\}) # examplesIf}
}
