% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/plot_implicit_forecasts.R,
%   R/plot_underlying_forecasts.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{implicit_forecasts}
\alias{underlying_forecasts}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2:autoplot]{autoplot()}}}

  \item{rjd3filters}{\code{\link[rjd3filters:implicit_forecasts]{implicit_forecasts()}}, \code{\link[rjd3filters:underlying_forecasts]{underlying_forecasts()}}}
}}

