% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_underlying_forecasts.R
\name{underlying_forecasts_plot}
\alias{underlying_forecasts_plot}
\alias{ggunderlying_forecasts_plot}
\title{Implicit Forecasts plot}
\usage{
underlying_forecasts_plot(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  col_i_f = col_sa,
  xlab = "",
  ylab = "",
  lty_last_tc = 2,
  lty_i_f = 3,
  n_last_tc = 4,
  ...
)

ggunderlying_forecasts_plot(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  col_i_f = col_sa,
  lty_last_tc = 2,
  lty_i_f = 3,
  n_last_tc = 4,
  legend_tc = "Trend-cycle",
  legend_sa = "Seasonally adjusted",
  legend_i_f = "Implicit forecasts",
  ...
)
}
\arguments{
\item{object}{\code{"tc_estimates"} object.}

\item{xlim, ylim}{x and y limits of the plot.
If \code{xlim} is defined and not \code{ylim}, then \code{ylim} is determined automatically.}

\item{col_sa, col_tc}{color of the seasonally adjusted and trend-cycle components.}

\item{col_i_f}{color of the underlying forecasts.}

\item{xlab, ylab}{x and y axis labels.}

\item{lty_last_tc, lty_i_f}{line type of the last values of the trend-cycle component and for the underlying forecasts.}

\item{n_last_tc}{number of last values of the trend-cycle component to be plotted with a different line type
(to emphasize that there is higher variability for the last estimates).
If \code{NULL}, then \code{n_last_tc} is equal to the MCD statistic.}

\item{...}{other parameters.}

\item{legend_tc, legend_sa, legend_i_f}{legend of the trend-cycle and seasonally adjusted components and for underlying forecasts.}
}
\value{
\code{underlying_forecasts_plot()} returns \code{NULL}, called for side effect.
\code{ggunderlying_forecasts_plot()} returns a \code{ggplot} object.
}
\description{
Implicit Forecasts plot
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
tc_mod <- henderson_smoothing(french_ipi[, "manufacturing"])
underlying_forecasts_plot(tc_mod, xlim = c(2022, 2025))
\dontshow{\}) # examplesIf}
}
