% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bandwidth}
\alias{bandwidth}
\title{Get Bandwidth}
\usage{
bandwidth(x)
}
\arguments{
\item{x}{a \code{"tc_estimates"} object.}
}
\description{
The bandwidth of a \code{"tc_estimates"} object is the number of observations on each side of symmetric
the trend-cycle filter used to compute the trend-cycle estimates.
The length of the filter is then equal to \eqn{2 \times \text{bandwidth}(x) + 1}.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
x <- french_ipi[, "manufacturing"]
tc_clf <- henderson_smoothing(x)
bandwidth(tc_clf)
\dontshow{\}) # examplesIf}
}
