% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bandwidth}
\alias{bandwidth}
\title{Get Bandwidth}
\usage{
bandwidth(x)
}
\arguments{
\item{x}{a \code{"tc_estimates"} object.}
}
\description{
Get the bandwidth of a \code{"tc_estimates"} object.
The length of the filter is then equal to \eqn{2 \times \text{bandwidth}(x) + 1}.
}
