% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classical-ma}
\alias{classical-ma}
\alias{henderson}
\alias{CLF}
\alias{CLF_CN}
\alias{local_param_est}
\title{Classical Moving Average}
\format{
An object of class \code{list} of length 5.

An object of class \code{finite_filters} of length 1.

An object of class \code{finite_filters} of length 1.

An object of class \code{list} of length 5.
}
\usage{
henderson

CLF

CLF_CN

local_param_est
}
\description{
Classical Moving Average
}
\keyword{datasets}
