% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint-tc}
\alias{confint-tc}
\alias{confint.henderson}
\alias{confint.clf}
\alias{confint.robust_henderson}
\title{Confidence Intervals for \code{"tc_estimates"}}
\usage{
\method{confint}{henderson}(object, parm, level = 0.95, ...)

\method{confint}{clf}(object, parm, level = 0.95, ...)

\method{confint}{robust_henderson}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a \code{"tc_estimates"} object.}

\item{parm}{unused parameter.}

\item{level}{the confidence level required.}

\item{...}{other (unused) parameters.}
}
\description{
Confidence Intervals for \code{"tc_estimates"}
}
