% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confint.R
\name{confint_plot}
\alias{confint_plot}
\alias{ggconfint_plot}
\title{Plot Confidence Intervals for Trend-Cycle Estimates}
\usage{
confint_plot(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  col_confint = "grey",
  xlab = "",
  ylab = "",
  level = 0.95,
  ...
)

ggconfint_plot(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  col_confint = "grey",
  legend_tc = "Trend-cycle",
  legend_sa = "Seasonally adjusted",
  legend_confint = "Confidence interval",
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{An object of class \code{"tc_estimates"}.}

\item{xlim, ylim}{Limits for the x- and y-axes.
If \code{xlim} is specified and \code{ylim} is \code{NULL}, \code{ylim} is determined automatically
based on the truncated series.}

\item{col_sa, col_tc}{Colors used for the seasonally adjusted and trend-cycle components,
respectively.}

\item{col_confint}{Color of the confidence interval lines or shaded region.}

\item{xlab, ylab}{Character strings for the x- and y-axis labels.}

\item{level}{The confidence level required (defaults to \code{0.95}).}

\item{...}{Additional arguments (currently unused).}

\item{legend_tc, legend_sa, legend_confint}{Character strings specifying the legend labels for
the trend-cycle, the seasonally adjusted series, and the confidence intervals, respectively.}
}
\value{
\itemize{
\item \code{confint_plot()} returns \code{NULL} invisibly and is called for its side effect (drawing a plot).
\item \code{ggconfint_plot()} returns a \code{\link[ggplot2]{ggplot}} object.
}
}
\description{
Base R (\code{confint_plot()}) and \pkg{ggplot2} (\code{ggconfint_plot()}) graphics for
plotting trend-cycle estimates alongside their confidence intervals.
}
\details{
\code{confint_plot()} produces a base R graphic, whereas \code{ggconfint_plot()} creates a
\pkg{ggplot2} object that can be further customized.
The confidence intervals are computed internally using \link[=confint_tc]{confint()}.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
tc_mod <- henderson_smoothing(french_ipi[, "manufacturing"])
confint_plot(tc_mod, xlim = c(2022, 2024.5))
\dontshow{\}) # examplesIf}
}
