% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confint.R
\name{confint_plot}
\alias{confint_plot}
\alias{ggconfint_plot}
\title{confint_plot plot}
\usage{
confint_plot(
  x,
  sa = NULL,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  col_confint = "grey",
  ...
)

ggconfint_plot(
  object,
  sa = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  col_confint = "grey",
  color_points = col_sa,
  ...
)
}
\description{
confint_plot plot
}
