% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confint.R
\name{confint_plot}
\alias{confint_plot}
\alias{ggconfint_plot}
\title{Confidence Intervals plot}
\usage{
confint_plot(
  object,
  sa = NULL,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  col_confint = "grey",
  xlab = "",
  ylab = "",
  ...
)

ggconfint_plot(
  object,
  sa = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  col_confint = "grey",
  legend_tc = "Trend-cycle",
  legend_sa = "Seasonally adjusted",
  legend_confint = "Confidence interval",
  ...
)
}
\arguments{
\item{object}{Data with trend-cycle component and the upper and lower bounds of the confidence interval.
If \code{object} is a \code{"tc_estimates"} then is is computed using the \code{confint()} function.}

\item{sa}{Seasonally adjusted components.
If \code{tc} is a \code{"tc_estimates"} object then \code{sa} is optional.}

\item{xlim, ylim}{x and y limits of the plot.
If \code{NULL} (the default), then the limits determined automatically.}

\item{col_sa, col_tc}{color of the seasonally adjusted and trend-cycle components.}

\item{col_confint}{color of the confidence interval.}

\item{xlab, ylab}{x and y axis labels.}

\item{...}{other parameters.}

\item{legend_tc, legend_sa, legend_confint}{legend of the trend-cycle and seasonally adjusted components and for the confidence intervals.}
}
\description{
Confidence Intervals plot
}
