% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confint.R
\name{confint_plot}
\alias{confint_plot}
\alias{ggconfint_plot}
\title{Confidence Intervals plot}
\usage{
confint_plot(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  col_confint = "grey",
  xlab = "",
  ylab = "",
  level = 0.95,
  ...
)

ggconfint_plot(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  col_confint = "grey",
  legend_tc = "Trend-cycle",
  legend_sa = "Seasonally adjusted",
  legend_confint = "Confidence interval",
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{\code{"tc_estimates"}.
The confidence intervals are computed using the \code{\link[=confint]{confint()}} function.}

\item{xlim, ylim}{x and y limits of the plot.
If \code{xlim} is defined and not \code{ylim}, then \code{ylim} is determined automatically.}

\item{col_sa, col_tc}{color of the seasonally adjusted and trend-cycle components.}

\item{col_confint}{color of the confidence interval.}

\item{xlab, ylab}{x and y axis labels.}

\item{level}{the confidence level required.}

\item{...}{other parameters.}

\item{legend_tc, legend_sa, legend_confint}{legend of the trend-cycle and seasonally adjusted components and for the confidence intervals.}
}
\description{
Confidence Intervals plot
}
\examples{
tc_mod <- henderson_smoothing(french_ipi[, "manufacturing"])
confint_plot(tc_mod, xlim = c(2022, 2024.5))
}
