% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{ggsmoothing_plot}
\alias{ggsmoothing_plot}
\title{Produce several plots}
\usage{
ggsmoothing_plot(
  object,
  plots = c("normal", "confint", "lollypop", "implicit_forecasts"),
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{\code{"tc_estimates"}.
The confidence intervals are computed using the \code{\link[=confint]{confint()}} function.}

\item{plots}{list of plots to use.}

\item{level}{the confidence level required.}

\item{...}{other unused parameters.}
}
\description{
Produce several plots
}
