% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{growthplot}
\alias{growthplot}
\alias{gggrowthplot}
\title{Plot Growth Rates for Trend-Cycle Estimates}
\usage{
growthplot(
  object,
  pct = TRUE,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  xlab = "",
  ylab = "",
  sa_bar_line = TRUE,
  ...,
  lag = -1
)

gggrowthplot(
  object,
  pct = TRUE,
  xlim = NULL,
  ylim = NULL,
  sa_bar_line = TRUE,
  col_tc = "#E69F00",
  col_sa = "black",
  col_sa_fill = "grey",
  legend_tc = "Trend-cycle",
  legend_sa = "Seasonally adjusted",
  ...,
  lag = -1
)
}
\arguments{
\item{object}{An object of class \code{"tc_estimates"}.}

\item{pct}{Logical. If \code{TRUE} (the default), the growth rate is expressed as a
percentage, otherwise as a ratio.}

\item{xlim, ylim}{Limits for the x- and y-axes.
If \code{xlim} is specified and \code{ylim} is \code{NULL}, \code{ylim} is determined automatically
based on the truncated series.}

\item{col_sa, col_tc}{Colors used for the seasonally adjusted and trend-cycle components,
respectively.}

\item{xlab, ylab}{Character strings for the x- and y-axis labels.}

\item{sa_bar_line}{Logical. If \code{TRUE} (the default), the growth rates of the
seasonally adjusted series are presented as bar-lines; otherwise, as lines.}

\item{...}{Additional graphical parameters passed to internal plotting functions.}

\item{lag}{Integer specifying the lag used to compute the growth rate.
Defaults to \code{-1} (i.e., period-to-period growth rate).}

\item{col_sa_fill}{Fill color of the bars for the seasonally adjusted series.}

\item{legend_tc, legend_sa}{Character strings specifying the legend labels for
the trend-cycle and seasonally adjusted components, respectively.}
}
\value{
\itemize{
\item \code{growthplot()} returns \code{NULL} invisibly and is called for its side effect (drawing a plot).
\item \code{gggrowthplot()} returns a \code{\link[ggplot2]{ggplot}} object.
}
}
\description{
Base R (\code{growthplot()}) and \pkg{ggplot2} (\code{gggrowthplot()}) graphics for
plotting the growth rates of the trend-cycle estimates (solid lines) and
the seasonally adjusted series (bar-lines or lines).
}
\details{
\code{growthplot()} produces a base R graphic, whereas \code{gggrowthplot()} creates a
\pkg{ggplot2} object that can be further customized.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
tc_mod <- henderson_smoothing(french_ipi[, "manufacturing"])
growthplot(tc_mod, xlim = c(2022, 2024.5))
\dontshow{\}) # examplesIf}
}
