% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{growthplot}
\alias{growthplot}
\alias{gggrowthplot}
\title{Growth plots}
\usage{
growthplot(
  object,
  pct = TRUE,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  xlab = "",
  ylab = "",
  sa_bar_line = TRUE,
  ...,
  lag = -1
)

gggrowthplot(
  object,
  pct = TRUE,
  xlim = NULL,
  ylim = NULL,
  sa_bar_line = TRUE,
  col_tc = "#E69F00",
  col_sa = "black",
  col_sa_fill = "grey",
  legend_tc = "Trend-cycle",
  legend_sa = "Seasonally adjusted",
  ...,
  lag = -1
)
}
\arguments{
\item{object}{\code{"tc_estimates"} object.}

\item{pct}{logical. If \code{TRUE} (the default), the growth rate is expressed in percentage points.}

\item{xlim, ylim}{x and y limits of the plot.
If \code{xlim} is defined and not \code{ylim}, then \code{ylim} is determined automatically.}

\item{col_sa, col_tc}{color of the seasonally adjusted and trend-cycle components.}

\item{xlab, ylab}{x and y axis labels.}

\item{sa_bar_line}{logical. If \code{TRUE} (the default), the growth rates of the seasonally adjusted series are
presented as bar-lines, otherwise they are presented as lines.}

\item{...}{other (unused) parameters.}

\item{lag}{lag used for the growth rate.
By default, \code{lag = -1} (i.e. period-to-period growth rate).}

\item{col_sa_fill}{fill color of the bar of the seasonally adjusted series.}

\item{legend_tc, legend_sa}{legend of the trend-cycle and seasonally adjusted components.}
}
\description{
Plots the growth rate of the trend-cycle (solid lines) and the seasonally adjusted series (bar-line).
}
\examples{
tc_mod <- henderson_smoothing(french_ipi[, "manufacturing"])
growthplot(tc_mod, xlim = c(2022, 2024.5))
}
