% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_filter.R
\name{henderson_robust_smoothing}
\alias{henderson_robust_smoothing}
\title{Smoothing using the Henderson filter}
\usage{
henderson_robust_smoothing(
  x,
  endpoints = c("Musgrave", "QL", "CQ", "DAF"),
  length = NULL,
  ao = NULL,
  ao_tc = NULL,
  ls = NULL,
  icr = NULL,
  local_icr = FALSE,
  asymmetric_var = FALSE,
  degree = 3,
  ...
)
}
\arguments{
\item{x}{input time-series.}

\item{endpoints}{Method used for the asymmetric filter.
By default the Musgrave method is used}

\item{length}{the length of the}

\item{ao}{Dates of the Additive Outliers (AO) which effects are associated to the irregular component.}

\item{ao_tc}{Dates of the Additive Outliers (AO) which effects are associated to the trend-cycle component.}

\item{ls}{Dates of the Level Shifts (LS) which effects are associated to the trend-cycle component.}

\item{icr}{I/C ratio used for the asymmetric filter.}

\item{local_icr}{if \code{TRUE}, the I/C ratio is estimated locally (as described in Quartier-la-Tente, A. (2024)) instead of globally.}

\item{asymmetric_var}{when \code{local_icr = TRUE},  if \code{asymmetric_var = TRUE} then the variance is estimated for each asymmetric filters instead of using the variance associated to the symmetric estimates.}

\item{degree}{if \code{local_icr = TRUE}, degree of polynomial used to estimate the local bias parameter.}

\item{...}{other parameters passed to \code{\link[rjd3filters:lp_filter]{rjd3filters::lp_filter()}}.}
}
\value{
An object of class \code{c("tc_estimates", "robust_henderson")}.
See \code{\link[=tc_estimates]{tc_estimates()}} for a full description of the returned object.
}
\description{
Smoothing using the Henderson filter
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
x <- cars_registrations
outliers <- x13_regarima_outliers(x)
tc <- henderson_robust_smoothing(x, ao = outliers$ao, ls = outliers$ls, local_icr = TRUE)

tc_henderson <- henderson_smoothing(x)
plot(tc, xlim = c(2019, 2022))
lines(tc_henderson$tc, col = "blue")
\dontshow{\}) # examplesIf}
}
