% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/henderson.R
\name{henderson_smoothing}
\alias{henderson_smoothing}
\title{Smoothing using the Henderson filter}
\usage{
henderson_smoothing(
  x,
  endpoints = c("Musgrave", "QL", "CQ", "CC", "DAF", "CN"),
  length = NULL,
  icr = NULL,
  local_icr = FALSE,
  local_var = TRUE,
  degree = 3,
  ...
)
}
\arguments{
\item{x}{input time-series.}

\item{endpoints}{Method used for the asymmetric filter.
By default the Musgrave method is used}

\item{length}{the length of the}

\item{icr}{I/C ratio used for the asymmetric filter.}

\item{local_icr}{if \code{TRUE}, the I/C ratio is estimated locally (as described in Quartier-la-Tente, A. (2024)) instead of globally.}

\item{local_var}{if \code{local_icr = TRUE} then the variance is estimated for each asymmetric filters instead of using the variance of the symmetric estimation.}

\item{degree}{if \code{local_icr = TRUE}, degree of polynomial used to estimate the local bias parameter.}

\item{...}{other parameters passed to \code{\link[rjd3filters:lp_filter]{rjd3filters::lp_filter()}}.}
}
\description{
Smoothing using the Henderson filter
}
\references{
Quartier-la-Tente, A. (2024). Improving Real-Time Trend Estimates Using Local Parametrization of Polynomial Regression Filters. \emph{Journal of Official Statistics, 40}(4), 685-715. \url{https://doi.org/10.1177/0282423X241283207}.
}
