% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr.R
\name{icr}
\alias{icr}
\title{Compute IC-Ratio}
\usage{
icr(x, sc, mul = FALSE)
}
\arguments{
\item{x}{input time series.}

\item{sc}{trend-cycle component.}

\item{mul}{boolean indicating if the decomposition is multiplicative or additive.}
}
\description{
Compute IC-Ratio
}
