% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_implicit_forecasts.R
\name{implicit_forecasts.henderson}
\alias{implicit_forecasts.henderson}
\alias{implicit_forecasts.clf}
\alias{implicit_forecasts.robust_henderson}
\title{Implicit Forecasts Methods}
\usage{
\method{implicit_forecasts}{henderson}(x, ...)

\method{implicit_forecasts}{clf}(x, ...)

\method{implicit_forecasts}{robust_henderson}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"tc_estimates"}.}

\item{...}{Additional arguments passed to internal methods (currently unused).}
}
\value{
A \code{matrix} or \code{mts} object containing the implicit forecasts for the series.
}
\description{
S3 method of \code{\link[rjd3filters:implicit_forecasts]{rjd3filters::implicit_forecasts()}} for \code{"tc_estimates"} objects.
}
\details{
See \code{\link[rjd3filters:implicit_forecasts]{rjd3filters::implicit_forecasts()}} for details on the computation of implicit forecasts.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
tc_mod <- henderson_smoothing(french_ipi[, "manufacturing"])
implicit_forecasts(tc_mod)
\dontshow{\}) # examplesIf}
}
