% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_implicit_forecast.R
\name{implicit_forecasts_plot}
\alias{implicit_forecasts_plot}
\alias{ggimplicit_forecasts_plot}
\title{Implicit Forecasts plot}
\usage{
implicit_forecasts_plot(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  col_i_f = col_sa,
  xlab = "",
  ylab = "",
  lty_last_tc = 2,
  lty_i_f = 3,
  ...
)

ggimplicit_forecasts_plot(
  object,
  col_tc = "#E69F00",
  col_sa = "black",
  col_i_f = col_sa,
  lty_last_tc = 2,
  lty_i_f = 3,
  legend_tc = "Trend-cycle",
  legend_sa = "Seasonally adjusted",
  legend_i_f = "Implicit forecasts",
  ...
)
}
\arguments{
\item{object}{\code{"tc_estimates"} object.
If \code{object} is a \code{"tc_estimates"} object then \code{sa} is optional.}

\item{xlim, ylim}{x and y limits of the plot.
If \code{NULL} (the default), then the limits determined automatically.}

\item{col_sa, col_tc}{color of the seasonally adjusted and trend-cycle components.}

\item{col_i_f}{color of the implicit forecasts.}

\item{xlab, ylab}{x and y axis labels.}

\item{lty_last_tc, lty_i_f}{line type of the last values of the trend-cycle component and for the implicit forecasts.}

\item{...}{other parameters.}

\item{legend_tc, legend_sa, legend_i_f}{legend of the trend-cycle and seasonally adjusted components and for implicit forecasts.}
}
\description{
Implicit Forecasts plot
}
