% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lollypop.R
\name{lollypop}
\alias{lollypop}
\alias{gglollypop}
\title{Lollipop Plot for Trend-Cycle Estimates}
\usage{
lollypop(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  color_points = col_sa,
  cex_points = 1,
  pch_points = 16,
  xlab = "",
  ylab = "",
  lty_last_tc = 2,
  n_last_tc = 4,
  ...
)

gglollypop(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  color_points = col_sa,
  cex_points = 1,
  pch_points = 16,
  legend_tc = "Trend-cycle",
  legend_sa = "Seasonally adjusted",
  lty_last_tc = 2,
  n_last_tc = 4,
  ...
)
}
\arguments{
\item{object}{An object of class \code{"tc_estimates"}.}

\item{xlim, ylim}{Limits for the x- and y-axes.
If \code{xlim} is specified and \code{ylim} is \code{NULL}, \code{ylim} is determined automatically
based on the truncated series.}

\item{col_sa, col_tc}{Colors used for the seasonally adjusted and trend-cycle components,
respectively.}

\item{color_points, cex_points}{Color and size (character/numeric) of the points
associated with the seasonally adjusted component.}

\item{pch_points}{Point symbol type (integer or character) for the seasonally adjusted component.}

\item{xlab, ylab}{Character strings for the x- and y-axis labels.}

\item{lty_last_tc}{Line type for the last values of the trend-cycle component.}

\item{n_last_tc}{Number of final values of the trend-cycle component to plot
with a distinct line type (\code{lty_last_tc}), emphasizing higher uncertainty in
recent estimates. If \code{NULL} (the default), \code{n_last_tc} is set to the Month of
Cyclical Dominance (MCD) statistic.}

\item{...}{Additional graphical parameters passed to internal plotting functions.}

\item{legend_tc, legend_sa}{Character strings specifying the legend labels for
the trend-cycle and seasonally adjusted components, respectively.}
}
\value{
\itemize{
\item \code{lollypop()} returns \code{NULL} invisibly and is called for its side effect (drawing a plot).
\item \code{gglollypop()} returns a \code{\link[ggplot2]{ggplot}} object.
}
}
\description{
Base R (\code{lollypop()}) and \pkg{ggplot2} (\code{gglollypop()}) graphics for creating
lollipop plots of trend-cycle estimates alongside seasonally adjusted series.
}
\details{
\code{lollypop()} produces a base R graphic, whereas \code{gglollypop()} creates a
\pkg{ggplot2} object that can be further customized.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
tc_mod <- henderson_smoothing(french_ipi[, "manufacturing"])
lollypop(tc_mod, xlim = c(2022, 2024.5))
\dontshow{\}) # examplesIf}
}
