% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollypop.R
\name{lollypop}
\alias{lollypop}
\alias{gglollypop}
\title{Lollypop plot}
\usage{
lollypop(
  tc,
  sa,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  color_points = col_sa,
  cex_points = 1,
  xlab = "",
  ylab = "",
  ...
)

gglollypop(
  tc,
  sa,
  col_tc = "#E69F00",
  col_sa = "black",
  color_points = col_sa,
  cex_points = 1,
  ...
)
}
\arguments{
\item{tc, sa}{Seasonally adjusted and trend-cycle components.
If \code{tc} is a \code{"tc_estimates"} object then \code{sa} is optional}

\item{xlim, ylim}{x and y limits of the plot.
If \code{NULL} (the default), then the limits determined automatically.}

\item{col_sa, col_tc}{color of the seasonally adjusted and trend-cycle components.}

\item{color_points, cex_points}{color and size of the points associated to the seasonnaly adjusted component.}
}
\description{
Lollypop plot
}
