% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr.R
\name{mcd}
\alias{mcd}
\title{Month of Cyclical Dominance}
\usage{
mcd(x, tc, mul = FALSE)
}
\arguments{
\item{x, tc}{seasonally adjusted and trend-cycle components.
If \code{x} is a \code{"tc_estimates"} object then \code{tc} is ignored.}

\item{mul}{boolean indicating if the decomposition is multiplicative or additive.}
}
\description{
Month of Cyclical Dominance
}
