% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tc_estimates}
\alias{plot.tc_estimates}
\alias{autoplot.tc_estimates}
\alias{autoplot.ts}
\title{Default \code{"tc_estimates"} plot}
\usage{
\method{plot}{tc_estimates}(
  x,
  y = NULL,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  xlab = "",
  ylab = "",
  lty_last_tc = 2,
  n_last_tc = 4,
  ...
)

\method{autoplot}{tc_estimates}(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  legend_tc = "Trend-cycle",
  legend_sa = "Seasonally adjusted",
  lty_last_tc = 2,
  n_last_tc = 4,
  ...
)

\method{autoplot}{ts}(object, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{y}{unused parameter.}

\item{xlim, ylim}{x and y limits of the plot.
If \code{xlim} is defined and not \code{ylim}, then \code{ylim} is determined automatically.}

\item{col_sa, col_tc}{color of the seasonally adjusted and trend-cycle components.}

\item{xlab, ylab}{x and y axis labels.}

\item{lty_last_tc}{line type of the last values of the trend-cycle component.}

\item{n_last_tc}{number of last values of the trend-cycle component to be plotted with a different line type
(to emphasize that there is higher variability for the last estimates).
If \code{NULL}, then \code{n_last_tc} is equal to the MCD statistic.}

\item{...}{other (unused) parameters.}

\item{object, x}{\code{"tc_estimates"} object.}

\item{legend_tc, legend_sa}{legend of the trend-cycle and seasonally adjusted components.}
}
\value{
\code{plot()} returns \code{NULL}, called for side effect.
\code{autoplot()} returns a \code{ggplot} object.
}
\description{
Default \code{"tc_estimates"} plot
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
tc_mod <- henderson_smoothing(french_ipi[, "manufacturing"])
plot(tc_mod, xlim = c(2022, 2024.5))
\dontshow{\}) # examplesIf}
}
