% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tc_estimates}
\alias{plot.tc_estimates}
\alias{autoplot.tc_estimates}
\title{Default \code{"tc_estimates"} plot}
\usage{
\method{plot}{tc_estimates}(
  x,
  y = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  xlab = "",
  ylab = "",
  lty_last_tc = 2,
  ...
)

\method{autoplot}{tc_estimates}(
  object,
  col_tc = "#E69F00",
  col_sa = "black",
  legend_tc = "Trend-cycle",
  legend_sa = "Seasonally adjusted",
  lty_last_tc = 2,
  ...
)
}
\arguments{
\item{y}{unused parameter.}

\item{col_sa, col_tc}{color of the seasonally adjusted and trend-cycle components.}

\item{xlab, ylab}{x and y axis labels.}

\item{lty_last_tc}{line type of the last values of the trend-cycle component.}

\item{...}{other (unused) parameters.}

\item{object, x}{\code{"tc_estimates"} object.}

\item{legend_tc, legend_sa}{legend of the trend-cycle and seasonally adjusted components.}
}
\description{
Default \code{"tc_estimates"} plot
}
