% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tc_estimates}
\alias{plot.tc_estimates}
\alias{autoplot.tc_estimates}
\title{Default \code{"tc_estimates"} plot}
\usage{
\method{plot}{tc_estimates}(
  x,
  y = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  xlab = "",
  ylab = "",
  ...
)

\method{autoplot}{tc_estimates}(object, col_tc = "#E69F00", col_sa = "black", ...)
}
\arguments{
\item{y}{unused parameter.}

\item{col_sa, col_tc}{color of the seasonally adjusted and trend-cycle components.}

\item{xlab, ylab}{x and y axis labels.}

\item{...}{other (unused) parameters.}

\item{object, x}{\code{"tc_estimates"} object.}
}
\description{
Default \code{"tc_estimates"} plot
}
