% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{smoothing}
\alias{smoothing}
\title{Smoothing using several methods}
\usage{
smoothing(
  x,
  methods = c("henderson", "henderson_localic", "henderson_robust",
    "henderson_robust_localic", "clf_cn", "clf_alf"),
  endpoints = "Musgrave",
  length = NULL,
  icr = NULL,
  asymmetric_var = FALSE,
  degree = 3,
  ao = NULL,
  ao_tc = NULL,
  ls = NULL,
  ...
)
}
\arguments{
\item{x}{input time-series.}

\item{methods}{list of methods to use.}

\item{endpoints}{Method used for the asymmetric filter.
By default the Musgrave method is used}

\item{length}{the length of the}

\item{icr}{I/C ratio used for the asymmetric filter.}

\item{asymmetric_var}{when \code{local_icr = TRUE},  if \code{asymmetric_var = TRUE} then the variance is estimated for each asymmetric filters instead of using the variance associated to the symmetric estimates.}

\item{degree}{if \code{local_icr = TRUE}, degree of polynomial used to estimate the local bias parameter.}

\item{ao}{Dates of the Additive Outliers (AO) which effects are associated to the irregular component.}

\item{ao_tc}{Dates of the Additive Outliers (AO) which effects are associated to the trend-cycle component.}

\item{ls}{Dates of the Level Shifts (LS) which effects are associated to the trend-cycle component.}

\item{...}{other unused parameters.}
}
\description{
Smoothing using several methods and returning a list of trend-cycle estimates.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
x <- french_ipi[, "manufacturing"]
outliers <- x13_regarima_outliers(x)
all_methods <- smoothing(x, ao = outliers$ao, ao_tc = outliers$ao_tc, ls = outliers$ls)
ggsmoothing_plot(all_methods)
\dontshow{\}) # examplesIf}
}
