% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smoothness}
\alias{smoothness}
\title{Measure Smoothness of Trend-Cycle Estimates}
\usage{
smoothness(x)
}
\arguments{
\item{x}{An object of class \code{"tc_estimates"}.}
}
\value{
A numeric value representing the smoothness index of the trend-cycle estimates.
}
\description{
Computes the smoothness ratio of trend-cycle estimates relative to the seasonally
adjusted series following Picard and Matthews (2016).
}
\details{
Smoothness is evaluated as the ratio of month-to-month growth rate variances:
\deqn{
100 \times \sqrt{
\frac{
\sum_t \left( \frac{TC_t - TC_{t-1}}{TC_{t-1}} \right)^2
}{
\sum_t \left( \frac{SA_t - SA_{t-1}}{SA_{t-1}} \right)^2
}
}
}{100 * sqrt( sum( ((TC_t - TC_{t-1}) / TC_{t-1})^2 ) / sum( ((SA_t - SA_{t-1}) / SA_{t-1})^2 ) )}

This metric quantifies the relative variability of the trend-cycle compared to
the seasonally adjusted series. A lower value indicates a smoother trend-cycle.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
x <- french_ipi[, "manufacturing"]
tc_clf <- clf_smoothing(x)
tc_henderson <- henderson_smoothing(x)
smoothness(tc_clf)
smoothness(tc_henderson)
\dontshow{\}) # examplesIf}
}
\references{
Picard, F., & Matthews, S. (2016). The Addition of Trend-Cycle Estimates
to Selected Publications at Statistics Canada. \emph{Proceedings of the Survey
Methods Section, Statistical Society of Canada (SSC) Annual Meeting}.
\url{https://ssc.ca/sites/default/files/imce/pdf/picard_ssc2016.pdf}
}
