% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smoothness}
\alias{smoothness}
\title{Get Smoothness of Trend-Cycle estimates}
\usage{
smoothness(x)
}
\arguments{
\item{x}{a \code{"tc_estimates"} object.}
}
\description{
Compute the smoothness of trend-cycle estimates using Picard and Matthews (2006) definition.
}
\details{
There are several definitions or criteria of smoothness. The measure of smoothness used here is defined as:
\deqn{
100 \times \sqrt{
\frac{
\sum_t [(TC_t - TC_{t-1})/TC_{t-1}]^2
}{
\sum_t [(SA_t - SA_{t-1})/SA_{t-1}]^2
}
}
}
This represents a measure of the month-to-month percentage change in the trend-cycle.
The smaller the smoothness number is, the smoother the trend-cycle.
}
\references{
Picard, Frédéric et Steve Matthews (2016).
\emph{The Addition of Trend-Cycle Estimates to Selected Publications at Statistics Canada}.
Proceedings of the Survey Methods Section, Statistical Society of Canada (SSC) Annual Meeting. <https://ssc.ca/sites/default/files/imce/pdf/p icard_ssc2016.pdf>.
}
