% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_estimates.R
\name{tc_estimates}
\alias{tc_estimates}
\alias{is_tc_estimates}
\alias{summary.tc_estimates}
\title{Trend-Cycle Estimates Class}
\usage{
tc_estimates(tc, sa, parameters = NULL, extra_class = NULL, ...)

is_tc_estimates(x)

\method{summary}{tc_estimates}(object, ...)
}
\arguments{
\item{tc}{the trend-cycle estimates.}

\item{sa}{the original time series (usually seasonnally adjusted).}

\item{parameters}{a list of parameters used to compute the trend-cycle estimates (for example the filter coefficients).}

\item{extra_class}{additional class to add to the output object.}

\item{...}{other unused parameters.}

\item{x, object}{a \code{"tc_estimates"} object.}
}
\value{
\code{tc_estimates()} returns an object of class \code{c("tc_estimates", extra_class)} which is a list with the following components:
\itemize{
\item \code{tc}: the trend-cycle estimates.
\item \code{x}: the original time series.
\item \code{parameters}: a list of parameters used to compute the trend-cycle estimates.
}

The \code{summary()} method for \code{tc_estimates} objects returns a list with the following components:
\itemize{
\item \verb{I/C ratio}: the I/C ratio of the trend-cycle estimates (see \code{\link[=icr]{icr()}}).
\item \verb{I/C ratios}: the I/C ratio per period of the trend-cycle estimates (see \code{\link[=icrs]{icrs()}}).
\item \code{MCD}: the Month of Cyclical Dominance (see \code{\link[=mcd]{mcd()}}).
\item \code{Length}: the length of the trend-cycle filter used for the final estimates
(i.e. the number of observations used to estimate the trend-cycle at the center of the series).
}
}
\description{
Trend-Cycle Estimates Class
}
