% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts-exemple}
\alias{ts-exemple}
\alias{cars_registrations}
\alias{french_ipi}
\alias{fred}
\alias{simulated_data}
\title{Data set examples}
\format{
An object of class \code{ts} of length 177.

An object of class \code{mts} (inherits from \code{ts}, \code{matrix}, \code{array}) with 416 rows and 3 columns.

An object of class \code{mts} (inherits from \code{ts}, \code{matrix}, \code{array}) with 766 rows and 2 columns.

An object of class \code{mts} (inherits from \code{ts}, \code{matrix}, \code{array}) with 84 rows and 6 columns.
}
\usage{
cars_registrations

french_ipi

fred

simulated_data
}
\description{
Data set examples
}
\keyword{datasets}
