% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_underlying_forecasts.R
\name{underlying_forecasts.henderson}
\alias{underlying_forecasts.henderson}
\alias{underlying_forecasts.clf}
\alias{underlying_forecasts.robust_henderson}
\title{Underlying Forecasts Methods}
\usage{
\method{underlying_forecasts}{henderson}(x, ...)

\method{underlying_forecasts}{clf}(x, ...)

\method{underlying_forecasts}{robust_henderson}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"tc_estimates"}.}

\item{...}{Additional arguments passed to internal methods (currently unused).}
}
\value{
A \code{matrix} or \code{mts} object containing the underlying forecasts for the series.
}
\description{
S3 method of \code{\link[rjd3filters:underlying_forecasts]{rjd3filters::underlying_forecasts()}} for \code{"tc_estimates"} objects.
}
\details{
See \code{\link[rjd3filters:underlying_forecasts]{rjd3filters::underlying_forecasts()}} for details on the computation of underlying forecasts.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
tc_mod <- henderson_smoothing(french_ipi[, "manufacturing"])
underlying_forecasts(tc_mod)
\dontshow{\}) # examplesIf}
}
