% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_underlying_forecasts.R
\name{underlying_forecasts_plot}
\alias{underlying_forecasts_plot}
\alias{ggunderlying_forecasts_plot}
\title{Plot Underlying Forecasts for Trend-Cycle Estimates}
\usage{
underlying_forecasts_plot(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  col_i_f = col_sa,
  xlab = "",
  ylab = "",
  lty_last_tc = 2,
  lty_i_f = 3,
  n_last_tc = 4,
  ...
)

ggunderlying_forecasts_plot(
  object,
  xlim = NULL,
  ylim = NULL,
  col_tc = "#E69F00",
  col_sa = "black",
  col_i_f = col_sa,
  lty_last_tc = 2,
  lty_i_f = 3,
  n_last_tc = 4,
  legend_tc = "Trend-cycle",
  legend_sa = "Seasonally adjusted",
  legend_i_f = "Implicit forecasts",
  ...
)
}
\arguments{
\item{object}{An object of class \code{"tc_estimates"}.}

\item{xlim, ylim}{Limits for the x- and y-axes.
If \code{xlim} is specified and \code{ylim} is \code{NULL}, \code{ylim} is determined automatically
based on the truncated series.}

\item{col_sa, col_tc}{Colors used for the seasonally adjusted and trend-cycle components,
respectively.}

\item{col_i_f}{Color of the forecasts lines.}

\item{xlab, ylab}{Character strings for the x- and y-axis labels.}

\item{lty_last_tc, lty_i_f}{Line types used for the last values of the trend-cycle
component and for the forecasts, respectively.}

\item{n_last_tc}{Number of final values of the trend-cycle component to plot
with a distinct line type (\code{lty_last_tc}), emphasizing higher uncertainty in
recent estimates. If \code{NULL} (the default), \code{n_last_tc} is set to the Month of
Cyclical Dominance (MCD) statistic.}

\item{...}{Additional graphical parameters passed to internal plotting functions.}

\item{legend_tc, legend_sa, legend_i_f}{Character strings specifying the legend labels
for the trend-cycle, the seasonally adjusted series, and the forecasts, respectively.}
}
\value{
\itemize{
\item \code{underlying_forecasts_plot()} returns \code{NULL} invisibly and is called for its side effect (drawing a plot).
\item \code{ggunderlying_forecasts_plot()} returns a \code{\link[ggplot2]{ggplot}} object.
}
}
\description{
Base R (\code{underlying_forecasts_plot()}) and \pkg{ggplot2} (\code{ggunderlying_forecasts_plot()})
graphics for plotting trend-cycle estimates alongside their underlying forecasts.
}
\details{
\code{underlying_forecasts_plot()} produces a base R graphic, whereas
\code{ggunderlying_forecasts_plot()} creates a \pkg{ggplot2} object that can be further customized.
}
\examples{
\dontshow{if (rjd3jars::check_java_version(silent = TRUE)) withAutoprint(\{ # examplesIf}
tc_mod <- henderson_smoothing(french_ipi[, "manufacturing"])
underlying_forecasts_plot(tc_mod, xlim = c(2022, 2025))
\dontshow{\}) # examplesIf}
}
