% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp.R
\name{write.ts}
\alias{write.ts}
\alias{read.ts}
\title{Export and Import time series object to/from CSV}
\usage{
write.ts(x, file)

read.ts(file, frequency = NULL, list = FALSE)
}
\arguments{
\item{x}{a time series object}

\item{file}{a character string giving the name of the file to write to.}

\item{frequency}{an integer giving the number of observations per unit of time.
By default it is guessed from the data.}

\item{list}{boolean, if \code{TRUE}, the function returns a list of time series objects.}
}
\description{
Export and Import time series object to/from CSV
}
