% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icr.R
\name{x11_trend_selection}
\alias{x11_trend_selection}
\alias{x11_trend_selection.default}
\alias{x11_trend_selection.ts}
\title{X-11 Selection of Trend Filter}
\usage{
x11_trend_selection(x, ...)

\method{x11_trend_selection}{default}(x, ..., freq)

\method{x11_trend_selection}{ts}(x, ...)
}
\arguments{
\item{x}{I/C ratio \code{\link[=icr]{icr()}} or a time series}

\item{...}{further arguments passed to or from other methods.}

\item{freq}{frequency of the time series used to compute the I/C ratio.}
}
\description{
X-11 Selection of Trend Filter
}
\details{
The following procedure is used in X-11 to select the length of the trend filter:
\enumerate{
\item Computes the I/C ratio, \eqn{icr} with an Henderson filter of length 13.
\item The length depends on the value or \eqn{icr}:
\itemize{
\item if \eqn{icr < 1} then the selected length is 9 for monthly data and 5 otherwise;
\item if \eqn{1 \leq icr < 3.5} then the selected length is \eqn{freq + 1} where \eqn{freq} is the frequency of data (12 for monthly data, 4 for quarterly data...).
\item if \eqn{icr \geq 3.5} then the selected length is 23 for monthly data and 7 otherwise.
}
}
}
