% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13_regarima_outliers.R
\name{x13_regarima_outliers}
\alias{x13_regarima_outliers}
\title{Outlier detection with RegARIMA Model}
\usage{
x13_regarima_outliers(
  y,
  order = c(0, 1, 1),
  mean = FALSE,
  ao = TRUE,
  ls = TRUE
)
}
\arguments{
\item{y}{A \code{ts} object.}

\item{order}{Orders of the ARIMA model.}

\item{mean}{Logical, if \code{TRUE} the model includes a constant term.}

\item{ao, ls}{Boolean to indicate if additive outliers (AO) or level shifts (LS) should be detected.}
}
\value{
A list with two elements: \code{ao} and \code{ls}, which are vectors of time points where the respective outliers were detected. If no outliers were detected, the corresponding element will be \code{NULL}.
}
\description{
Wrapper around \code{\link[rjd3x13:regarima_outliers]{rjd3x13::regarima_outliers()}} to detect Additive Outliers (AO) or
Level Shifts (LS) in a seasonnally adjusted series.
}
